+++ assets/css/searchpage.css
@@ -1,92 +1,105 @@
 /* search-page */
+
 .searchDisplay{
     display: block;
     gap: 20px;
-    margin-top: 50px;
-    padding-top: 20px;
-    background-color: #292929;
+    width: 100%;
+    padding-block: 100px;
+    background-color: #0e0e0e;
+    cursor: pointer;
 }
 .search-block{
     border: none;
     background-color: rgb(102, 102, 190);
     display: inline-flex;
-    width: 200px;
+    width: 250px;
     position: relative;
-    height: 250px;
-    margin-inline: 30px;
+    height: 260px;
+    margin-inline: 20px;
     margin-block: 50px;
     overflow: hidden;
     border-radius: 10px;
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.2); 
+
 }
 .searchDisplay h1{
     color: white;
+    font-size: 30px;
+    font-weight: bolder;
+    padding-left: 20px;
 }
 .search-block h2{
     padding-left: 20px;
-    padding-top: 20px;
+    padding-top: 15px;
     font-weight: bolder;
     font-size: 25px;
+    color: white;
   }
 .search-block img {
-    width: 80%;
+    width: 60%;
     height: 60%;
     position: absolute;
-    bottom: -10%;
-    right:-20%;
-    transform: rotate(10deg);
+    bottom: -5%;
+    right:-8%;
+    transform: rotate(20deg);
   }
 .search-block:hover img{
     transform: scale(1.05);
 }
+
 /* search-bar */
-.search{
+
+
+.search-input{
     background-color: #121212;
-  }
-.search input[type="text"] {
-    padding-block: 20px;
-    padding-right: 100px;
+}
+.search-input[type="text"] {
     border: none;
     border-radius: 50px;
     background-color: #242424;;
     color: #b3b3b3;
     width: 300px;
     text-align: center;
-    outline: none;
-    color: #fff;
   }
-.search input[type="text"]::placeholder {
+  .search-input[type="text"]:onClick {
+    border: 50px solid white; 
+    border-radius: 50px; 
+  }
+  .search-input[type="text"]:hover{
+    background-color: #383737;
+  }
+.search-input[type="text"]::placeholder {
     color: #b3b3b3;
-    opacity: 1;
     font-weight: 400;
-    font-size: 18px;
+    font-size: 12px;
     padding-inline: 50px;
+    margin-left: 2px;
   }
 .search-input{
     background: #fff url('/assets/imgs/search-icon..png') no-repeat 10px center;
 }
 
 /* Media player*/
+
 .spotify-player {
     background-color: #000000;
-    padding: 20px;
     display: flex;
     align-items: center;
-    height: 10%;
 }
 .cover-image{
     display: flex;
+    padding-left: 10px;
+    padding-top: 10px;
 }
 .cover-image img {
-    max-width: 75px;
-    max-height: 75px;
+    max-width: 100px;
+    max-height: 100px;
     margin-right: 20px;
 }
 .cover-image li{
     list-style: none;
     color: white;
-    padding-bottom: 10px;
-  }
+}
 .cover-image ul{
     padding-block: 15%;
 }
@@ -94,10 +107,9 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    position: absolute;
-    left: 30%;
-    bottom: 1%;
-    
+    position: relative;
+    left: 15%;
+    padding-bottom: 10px;
 }
 .control-button {
     background-color: black;
